# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seed Server",
    "version": "0.5"
  },
  "paths": {
...
    "/v1/lessons/generate": {
      "post": {
        "summary": "Generate Lesson",
        "description": "Generate a new language learning lesson using LLM.\n\nReturns structured lesson with tasks, grading rules, and metadata.\nUses persona system for tone consistency.",
        "operationId": "generate_lesson_v1_lessons_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lessons/submit": {
      "post": {
        "summary": "Submit Lesson Answer",
        "description": "Submit an answer for a lesson task and get graded feedback.\n\nReturns grade result with feedback, or lesson summary if last task.",
        "operationId": "submit_lesson_answer_v1_lessons_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lessons": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "List Lessons",
        "description": "List all lessons for the authenticated user.",
        "operationId": "list_lessons_v1_lessons_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "Get Lesson",
        "description": "Get a specific lesson with attempts.",
        "operationId": "get_lesson_v1_lessons__lesson_id__get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lessons"
        ],
        "summary": "Delete Lesson Endpoint",
        "description": "Delete a lesson and its attempts.",
        "operationId": "delete_lesson_endpoint_v1_lessons__lesson_id__delete",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GradeResponse": {
        "properties": {
          "grade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GradeResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LessonSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "personaIdUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaidused"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          }
        },
        "type": "object",
        "title": "GradeResponse",
        "description": "Response after grading a submission."
      },
      "GradeResult": {
        "properties": {
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "correctAnswer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correctanswer"
          }
        },
        "type": "object",
        "required": [
          "taskId",
          "correct",
          "score",
          "feedback"
        ],
        "title": "GradeResult",
        "description": "Result of grading a single task submission."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Lesson": {
        "properties": {
          "lessonId": {
            "type": "string",
            "title": "Lessonid"
          },
          "mode": {
            "$ref": "#/components/schemas/LessonMode"
          },
          "targetLang": {
            "type": "string",
            "title": "Targetlang"
          },
          "nativeLang": {
            "type": "string",
            "title": "Nativelang"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "lessonId",
          "mode",
          "targetLang",
          "nativeLang",
          "level",
          "title",
          "tasks"
        ],
        "title": "Lesson",
        "description": "Complete lesson structure."
      },
...
      "LessonDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "lessonId": {
            "type": "string",
            "title": "Lessonid"
          }
        },
        "type": "object",
        "required": [
          "deleted",
          "lessonId"
        ],
        "title": "LessonDeleteResponse",
        "description": "Response for DELETE /v1/lessons/{lessonId}."
      },
      "LessonGenerateRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/LessonMode"
          },
          "targetLang": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Targetlang"
          },
          "nativeLang": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Nativelang"
          },
          "level": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "lessonLength": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 3.0,
            "title": "Lessonlength"
          },
          "personaId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaid"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "targetLang",
          "nativeLang",
          "level",
          "lessonLength"
        ],
        "title": "LessonGenerateRequest",
        "description": "Request to generate a new lesson."
      },
      "LessonGetResponse": {
        "properties": {
          "lesson": {
            "$ref": "#/components/schemas/Lesson"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/LessonAttemptInfo"
            },
            "type": "array",
            "title": "Attempts"
          },
          "totalAttempts": {
            "type": "integer",
            "title": "Totalattempts"
          },
          "completedCount": {
            "type": "integer",
            "title": "Completedcount"
          },
          "totalScore": {
            "type": "integer",
            "title": "Totalscore"
          },
          "personaIdUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaidused"
          }
        },
        "type": "object",
        "required": [
          "lesson",
          "attempts",
          "totalAttempts",
          "completedCount",
          "totalScore"
        ],
        "title": "LessonGetResponse",
        "description": "Response for GET /v1/lessons/{lessonId}."
      },
...
      "LessonListResponse": {
        "properties": {
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonListItem"
            },
            "type": "array",
            "title": "Lessons"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "lessons",
          "total"
        ],
        "title": "LessonListResponse",
        "description": "Response for GET /v1/lessons."
      },
      "LessonMode": {
        "type": "string",
        "enum": [
          "vocabulary",
          "grammar",
          "mixed"
        ],
        "title": "LessonMode",
        "description": "Learning mode for lesson generation."
      },
      "LessonResponse": {
        "properties": {
          "lesson": {
            "$ref": "#/components/schemas/Lesson"
          },
          "personaIdUsed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaidused"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          }
        },
        "type": "object",
        "required": [
          "lesson"
        ],
        "title": "LessonResponse",
        "description": "Response containing generated lesson."
      },
...
      "LessonSubmitRequest": {
        "properties": {
          "lessonId": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Lessonid"
          },
          "taskId": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Taskid"
          },
          "userAnswer": {
            "type": "string",
            "maxLength": 1000,
            "title": "Useranswer"
          },
          "personaId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Personaid"
          }
        },
        "type": "object",
        "required": [
          "lessonId",
          "taskId",
          "userAnswer"
        ],
        "title": "LessonSubmitRequest",
        "description": "Request to submit an answer for a task."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}